+++ src/app.enums.ts
@@ -1,0 +1,5 @@
+export enum APPSTATE {
+	init = 'init',
+	errorInit = 'errorInit',
+	ready = 'ready'
+}