+++ src/lib/components/Icon/Icon.tsx
@@ -1,17 +1,17 @@
 import React from 'react';
+import styles from './Icon.module.scss';
 
-export interface ISvgIconProp extends React.DetailedHTMLProps<React.ButtonHTMLAttributes<HTMLDivElement>, HTMLDivElement> {
-	// svg: string;
+export interface IIconProp extends React.DetailedHTMLProps<React.ButtonHTMLAttributes<HTMLDivElement>, HTMLDivElement> {
 	children: React.ReactNode;
 }
 
-export const SvgIcon = (props: ISvgIconProp) => {
+export const Icon = (props: IIconProp) => {
 
 	const { children, className = '', ...rest } = props;
 
 	const getCssClasses = () => {
 		const cssClasses: Array<string> = [];
-		cssClasses.push(`svg-icon`);
+		cssClasses.push(styles.icon);
 		cssClasses.push(className);
 		return cssClasses.filter(css => css).join(' ');
 	}