+++ src/showcase/examples/index.ts
@@ -23,7 +23,7 @@ export * from './TabsetPage';
 export * from './TooltipPage';
 export * from './TreeViewPage';
 export * from './TypographyPage';
-export * from './SvgIconPage';
+export * from './IconPage';
 export * from './SelectPage';
 
 export * from '../../showcase/assets';